+++ src/factory/error/alreadyInUse.spec.ts
@@ -17,7 +17,7 @@ describe('new AlreadyInUseError()', () => {
         assert.equal(error.entityName, entityName);
         assert.equal(error.fieldNames, fieldNames);
         assert.equal(error.message, message);
-        assert.equal(error.name, 'TTTSError');
+        assert.equal(error.name, 'SmartTheaterError');
         assert.equal(typeof error.stack, 'string');
     });
 
@@ -28,7 +28,7 @@ describe('new AlreadyInUseError()', () => {
         assert(error instanceof Error);
         assert.equal(error.entityName, entityName);
         assert.equal(error.fieldNames, fieldNames);
-        assert.equal(error.name, 'TTTSError');
+        assert.equal(error.name, 'SmartTheaterError');
         assert.equal(typeof error.message, 'string');
         assert.equal(typeof error.stack, 'string');
     });