+++ src/factory/error/notFound.spec.ts
@@ -15,7 +15,7 @@ describe('new NotFoundError()', () => {
         assert(error instanceof Error);
         assert.equal(error.entityName, entityName);
         assert.equal(error.message, message);
-        assert.equal(error.name, 'TTTSError');
+        assert.equal(error.name, 'SmartTheaterError');
         assert.equal(typeof error.stack, 'string');
     });
 
@@ -24,7 +24,7 @@ describe('new NotFoundError()', () => {
         const error = new NotFoundError(entityName);
         assert(error instanceof Error);
         assert.equal(error.entityName, entityName);
-        assert.equal(error.name, 'TTTSError');
+        assert.equal(error.name, 'SmartTheaterError');
         assert.equal(typeof error.message, 'string');
         assert.equal(typeof error.stack, 'string');
     });