+++ src/factory/error/notImplemented.spec.ts
@@ -13,14 +13,14 @@ describe('new NotImplementedError()', () => {
         const error = new NotImplementedError(message);
         assert(error instanceof Error);
         assert.equal(error.message, message);
-        assert.equal(error.name, 'TTTSError');
+        assert.equal(error.name, 'SmartTheaterError');
         assert.equal(typeof error.stack, 'string');
     });
 
     it('メッセージを指定しなくても、正しくインスタンス化できる', () => {
         const error = new NotImplementedError();
         assert(error instanceof Error);
-        assert.equal(error.name, 'TTTSError');
+        assert.equal(error.name, 'SmartTheaterError');
         assert.equal(typeof error.message, 'string');
         assert.equal(typeof error.stack, 'string');
     });