+++ src/factory/error/rateLimitExceeded.spec.ts
@@ -13,14 +13,14 @@ describe('new RateLimitExceededError()', () => {
         const error = new RateLimitExceededError(message);
         assert(error instanceof Error);
         assert.equal(error.message, message);
-        assert.equal(error.name, 'TTTSError');
+        assert.equal(error.name, 'SmartTheaterError');
         assert.equal(typeof error.stack, 'string');
     });
 
     it('メッセージを指定しなくても、正しくインスタンス化できる', () => {
         const error = new RateLimitExceededError();
         assert(error instanceof Error);
-        assert.equal(error.name, 'TTTSError');
+        assert.equal(error.name, 'SmartTheaterError');
         assert.equal(typeof error.message, 'string');
         assert.equal(typeof error.stack, 'string');
     });